requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
requests
rarfile
py7zr
patool
colorama
pyyaml